@@ -30,43 +30,102 @@ packages = kraken
 
 [entry_points]
 console_scripts =
-    kraken = kraken.kraken:cli
-    ketos = kraken.ketos:cli
+  kraken = kraken.kraken:cli
+  ketos = kraken.ketos:cli
+kraken.cli =
+  binarize = kraken.kraken:binarize
+  segment = kraken.kraken:segment
+  ocr = kraken.kraken:ocr
+  show = kraken.kraken:show
+  list = kraken.kraken:list_models
+  get = kraken.kraken:get
+ketos.cli = 
+  compile = kraken.ketos.dataset:compile
+  pretrain = kraken.ketos.pretrain:pretrain
+  train = kraken.ketos.recognition:train
+  test = kraken.ketos.recognition:test
+  publish = kraken.ketos.repo:publish
+  rotrain = kraken.ketos.ro:rotrain
+  roadd = kraken.ketos.ro:roadd
+  segtrain = kraken.ketos.segmentation:segtrain
+  segtest = kraken.ketos.segmentation:segtest
+  convert = kraken.ketos.weights:convert
+kraken.models = 
+  TorchVGSLModel = kraken.lib.vgsl:TorchVGSLModel
+  Wav2Vec2Mask = kraken.lib.pretrain:Wav2Vec2Mask
+  ROMLP = kraken.lib.ro:ROMLP
+kraken.lightning_modules =
+  blla = kraken.train:BLLASegmentationModel
+  vgsl = kraken.train:VGSLRecognitionModel
+  pretrain = kraken.lib.pretrain:RecognitionPretrainModel
+  ro = kraken.lib.ro:ROModel
+kraken.loaders =
+  safetensors = kraken.models:load_safetensors 
+  coreml = kraken.models:load_coreml
+kraken.writers =
+  safetensors = kraken.models:write_safetensors 
+  coreml = kraken.models:write_coreml
+kraken.configs =
+  Config = kraken.configs.base:Config
+  TrainingConfig = kraken.configs.base:TrainingConfig
+  TrainingDataConfig = kraken.configs.base:TrainingDataConfig
+  RecognitionTrainingDataConfig = kraken.configs.base:RecognitionTrainingDataConfig
+  SegmentationTrainingDataConfig = kraken.configs.base:SegmentationTrainingDataConfig
+  BLLASegmentationTrainingDataConfig = kraken.configs.vgsl:BLLASegmentationTrainingDataConfig
+  BLLASegmentationTrainingConfig = kraken.configs.vgsl:BLLASegmentationTrainingConfig
+  VGSLRecognitionTrainingDataConfig = kraken.configs.vgsl:VGSLRecognitionTrainingDataConfig
+  VGSLRecognitionTrainingConfig = kraken.configs.vgsl:VGSLRecognitionTrainingConfig
+  ROTrainingDataConfig = kraken.configs.ro:ROTrainingDataConfig
+  ROTrainingConfig = kraken.configs.ro:ROTrainingConfig
+  VGSLPreTrainingDataConfig = kraken.configs.pretrain:VGSLPreTrainingDataConfig
+  VGSLPreTrainingConfig = kraken.configs.pretrain:VGSLPreTrainingConfig
+kraken.tasks =
+  segmentation = kraken.tasks:SegmentationTaskModel
+  recognition = kraken.tasks:RecognitionTaskModel
+  alignment = kraken.tasks:ForcedAlignmentTaskModel
 
 [flake8]
 max_line_length = 160
 exclude = tests/*
 
 [options]
-python_requires = >=3.9,<3.13
+python_requires = >=3.10,<3.14
 install_requires =
     jsonschema
     lxml
     requests
     click>=8.1,<8.3
-    numpy~=2.0.0
+    numpy
     Pillow>=9.2.0
     regex
-    scipy~=1.13.0
+    scipy~=1.15.3
     protobuf>=3.0.0
-    coremltools~=8.1
-    jinja2~=3.0
-    python-bidi~=0.6.0
+    coremltools~=9.0
+    jinja2~=3.1.6
+    python-bidi~=0.6.6
     torchvision>=0.5.0
-    torch>=2.4.0,<=2.9
-    scikit-learn~=1.5.0
-    scikit-image~=0.24.0
-    shapely>=2.0.6,~=2.0.6
+    torch>=2.4.0,<=2.10
+    scikit-learn~=1.7.2
+    scikit-image~=0.25.2
+    shapely~=2.1.2
     pyarrow
     htrmopo>=0.3,~=0.3
-    lightning~=2.4.0
+    lightning~=2.6.0
     torchmetrics>=1.1.0
-    threadpoolctl~=3.5.0
+    threadpoolctl~=3.6.0
     platformdirs
     rich<14.1.0
+    safetensors
     iso639-lang
+    packaging
 
 [options.extras_require]
 test = hocr-spec; pytest
 pdf = pyvips
-augment = albumentations
+augment = torchvision>=0.5.0
+docs =
+    sphinx
+    sphinx-autoapi
+    sphinx-multiversion
+    sphinx-github-changelog
+    sphinx-autodoc-typehints
