@@ -16,24 +16,26 @@
 kraken.blla
 ~~~~~~~~~~~
 
-Trainable layout analysis tools for kraken for line and region detection. The
-line recognizer uses the baseline paradigm.
+Legacy API for inference using the trainable baseline and region detection
+segmenter. New code should use the SegmentationTaskModel from kraken.tasks.
 """
-
-import logging
-import uuid
-from typing import Any, Callable, Dict, List, Literal, Optional, Union
-
-import numpy as np
 import PIL
+import uuid
+import logging
+import warnings
+import numpy as np
 import shapely.geometry as geom
 import torch
 import torch.nn.functional as F
 import torchvision.transforms as tf
 
+from dataclasses import replace
+from torchvision.transforms import v2
 from importlib import resources
 from scipy.ndimage import gaussian_filter
 from skimage.filters import sobel
+from typing import Any, Callable, Literal, Optional, Union, TYPE_CHECKING
+
 
 from kraken.containers import BaselineLine, Region, Segmentation
 from kraken.lib import dataset, vgsl
@@ -48,14 +50,17 @@ from kraken.lib.util import get_im_str, is_bitonal
 
 __all__ = ['segment']
 
+if TYPE_CHECKING:
+    from PIL import Image
+
 logger = logging.getLogger(__name__)
 
 
-def compute_segmentation_map(im: PIL.Image.Image,
+def compute_segmentation_map(im: 'Image.Image',
                              mask: Optional[np.ndarray] = None,
                              model: vgsl.TorchVGSLModel = None,
                              device: str = 'cpu',
-                             autocast: bool = False) -> Dict[str, Any]:
+                             autocast: bool = False) -> dict[str, Any]:
     """
     Args:
         im: Input image
@@ -68,7 +73,7 @@ def compute_segmentation_map(im: PIL.Image.Image,
 
     Returns:
         A dictionary containing the heatmaps ('heatmap', torch.Tensor), class
-        map ('cls_map', Dict[str, Dict[str, int]]), the bounding regions for
+        map ('cls_map', dict[str, dict[str, int]]), the bounding regions for
         polygonization purposes ('bounding_regions', List[str]), the scale
         between the input image and the network output ('scale', float), and
         the scaled input image to the network ('scal_im', PIL.Image.Image).
@@ -94,7 +99,7 @@ def compute_segmentation_map(im: PIL.Image.Image,
         padding = (padding[0], padding[0], padding[1], padding[1])
 
     transforms = dataset.ImageInputTransforms(batch, height, width, channels, padding, valid_norm=False)
-    tf_idx, _ = next(filter(lambda x: isinstance(x[1], tf.ToTensor), enumerate(transforms.transforms)))
+    tf_idx, _ = next(filter(lambda x: isinstance(x[1], v2.PILToTensor), enumerate(transforms.transforms)))
     res_tf = tf.Compose(transforms.transforms[:tf_idx])
     scal_im = np.array(res_tf(im).convert('L'))
 
@@ -135,14 +140,14 @@ def compute_segmentation_map(im: PIL.Image.Image,
             'scal_im': scal_im}
 
 
-def vec_regions(heatmap: torch.Tensor, cls_map: Dict, scale: float, **kwargs) -> Dict[str, List[Region]]:
+def vec_regions(heatmap: torch.Tensor, cls_map: dict, scale: float, **kwargs) -> dict[str, list[Region]]:
     """
     Computes regions from a stack of heatmaps, a class mapping, and scaling
     factor.
 
     Args:
         heatmap: A stack of heatmaps of shape `NxHxW` output from the network.
-        cls_map: Dictionary mapping string identifiers to indices on the stack
+        cls_map: dictionary mapping string identifiers to indices on the stack
                  of heatmaps.
         scale: Scaling factor between heatmap and unscaled input image.
 
@@ -161,22 +166,22 @@ def vec_regions(heatmap: torch.Tensor, cls_map: Dict, scale: float, **kwargs) ->
 
 
 def vec_lines(heatmap: torch.Tensor,
-              cls_map: Dict[str, Dict[str, int]],
+              cls_map: dict[str, dict[str, int]],
               scale: float,
               text_direction: str = 'horizontal-lr',
-              regions: List[np.ndarray] = None,
+              regions: list[np.ndarray] = None,
               scal_im: np.ndarray = None,
-              suppl_obj: List[np.ndarray] = None,
+              suppl_obj: list[np.ndarray] = None,
               topline: Optional[bool] = False,
               raise_on_error: bool = False,
-              **kwargs) -> List[Dict[str, Any]]:
+              **kwargs) -> list[dict[str, Any]]:
     r"""
     Computes lines from a stack of heatmaps, a class mapping, and scaling
     factor.
 
     Args:
         heatmap: A stack of heatmaps of shape `NxHxW` output from the network.
-        cls_map: Dictionary mapping string identifiers to indices on the stack
+        cls_map: dictionary mapping string identifiers to indices on the stack
                  of heatmaps.
         scale: Scaling factor between heatmap and unscaled input image.
         text_direction: Text directions used as hints in the reading order
@@ -246,7 +251,7 @@ def segment(im: PIL.Image.Image,
             text_direction: Literal['horizontal-lr', 'horizontal-rl', 'vertical-lr', 'vertical-rl'] = 'horizontal-lr',
             mask: Optional[np.ndarray] = None,
             reading_order_fn: Callable = polygonal_reading_order,
-            model: Union[List[vgsl.TorchVGSLModel], vgsl.TorchVGSLModel] = None,
+            model: Union[list[vgsl.TorchVGSLModel], vgsl.TorchVGSLModel] = None,
             device: str = 'cpu',
             raise_on_error: bool = False,
             autocast: bool = False) -> Segmentation:
@@ -297,6 +302,9 @@ def segment(im: PIL.Image.Image,
         are high. In addition, neural reading order determination is disabled
         when more than one model outputs lines.
     """
+    warnings.warn('`blla.segment)` is deprecated and will be removed with kraken 8. Use `SegmentationTaskModel` instead.',
+                  DeprecationWarning)
+
     if model is None:
         logger.info('No segmentation model given. Loading default model.')
         model = vgsl.TorchVGSLModel.load_model(resources.files('kraken').joinpath('blla.mlmodel'))
@@ -305,7 +313,7 @@ def segment(im: PIL.Image.Image,
         model = [model]
 
     for nn in model:
-        if nn.model_type != 'segmentation':
+        if 'segmentation' not in nn.model_type:
             raise KrakenInvalidModelException(f'Invalid model type {nn.model_type} for {nn}')
         if 'class_mapping' not in nn.user_metadata:
             raise KrakenInvalidModelException(f'Segmentation model {nn} does not contain valid class mapping')
@@ -317,8 +325,7 @@ def segment(im: PIL.Image.Image,
     order = None
     regions = {}
     multi_lines = False
-    # flag to indicate that multiple models produced line output -> disable
-    # neural reading order
+
     for net in model:
         if 'topline' in net.user_metadata:
             loc = {None: 'center',
@@ -327,7 +334,7 @@ def segment(im: PIL.Image.Image,
             logger.debug(f'Baseline location: {loc}')
         rets = compute_segmentation_map(im, mask, net, device, autocast=autocast)
         _regions = vec_regions(**rets)
-        for reg_key, reg_val in vec_regions(**rets).items():
+        for reg_key, reg_val in _regions.items():
             if reg_key not in regions:
                 regions[reg_key] = []
             regions[reg_key].extend(reg_val)
@@ -350,14 +357,92 @@ def segment(im: PIL.Image.Image,
                            topline=net.user_metadata['topline'] if 'topline' in net.user_metadata else False,
                            raise_on_error=raise_on_error)
 
-        if 'ro_model' in net.aux_layers:
-            logger.info(f'Using reading order model found in segmentation model {net}.')
-            _order = neural_reading_order(lines=_lines,
-                                          regions=_regions,
-                                          text_direction=text_direction[-2:],
-                                          model=net.aux_layers['ro_model'],
-                                          im_size=im.size,
-                                          class_mapping=net.user_metadata['ro_class_mapping'])
+        # convert dicts to BaselineLine objects
+        _lines = [BaselineLine(id=f'_{uuid.uuid4()}',
+                               baseline=line['baseline'],
+                               boundary=line['boundary'],
+                               tags=line['tags']) for line in _lines]
+
+        if 'ro_model' in net.aux_layers or 'ro_model_regions' in net.aux_layers:
+            logger.info(f'Using reading order model(s) found in segmentation model {net}.')
+            flat_regs = [reg for rgs in _regions.values() for reg in rgs]
+            line_ro = net.aux_layers.get('ro_model')
+            region_ro = net.aux_layers.get('ro_model_regions')
+
+            # Region ordering
+            if region_ro and flat_regs:
+                reg_order = neural_reading_order(lines=flat_regs,
+                                                 model=region_ro,
+                                                 im_size=im.size,
+                                                 class_mapping=net.user_metadata['class_mapping']['regions'])
+                if reg_order is not None:
+                    ordered_regs = [flat_regs[i] for i in reg_order]
+                else:
+                    ordered_regs = flat_regs
+            else:
+                ordered_regs = flat_regs
+
+            # Line ordering (per-region if region model present, else global)
+            if line_ro:
+                line_class_mapping = net.user_metadata['class_mapping']['baselines']
+                if region_ro and ordered_regs:
+                    from collections import defaultdict
+                    region_line_map = defaultdict(list)
+                    region_ids = {reg.id for reg in ordered_regs}
+                    for line in _lines:
+                        if line.regions and line.regions[0] in region_ids:
+                            region_line_map[line.regions[0]].append(line)
+                        else:
+                            region_line_map[None].append(line)
+                    ordered_lines = []
+                    for reg in ordered_regs:
+                        rlines = region_line_map.get(reg.id, [])
+                        if len(rlines) > 1:
+                            lo = neural_reading_order(lines=rlines,
+                                                      model=line_ro,
+                                                      im_size=im.size,
+                                                      class_mapping=line_class_mapping)
+                            if lo is not None:
+                                ordered_lines.extend([rlines[i] for i in lo])
+                            else:
+                                ordered_lines.extend(rlines)
+                        else:
+                            ordered_lines.extend(rlines)
+                    orphans = region_line_map.get(None, [])
+                    if len(orphans) > 1:
+                        lo = neural_reading_order(lines=orphans,
+                                                  model=line_ro,
+                                                  im_size=im.size,
+                                                  class_mapping=line_class_mapping)
+                        if lo is not None:
+                            ordered_lines.extend([orphans[i] for i in lo])
+                        else:
+                            ordered_lines.extend(orphans)
+                    else:
+                        ordered_lines.extend(orphans)
+                    _order = [_lines.index(l) for l in ordered_lines]
+                else:
+                    _order = neural_reading_order(lines=_lines,
+                                                  regions=flat_regs,
+                                                  text_direction=text_direction[-2:],
+                                                  model=line_ro,
+                                                  im_size=im.size,
+                                                  class_mapping=line_class_mapping)
+            elif region_ro:
+                # Only region model — group lines by region order
+                ordered_lines = []
+                used = set()
+                for reg in ordered_regs:
+                    for line in _lines:
+                        if line.regions and line.regions[0] == reg.id and id(line) not in used:
+                            ordered_lines.append(line)
+                            used.add(id(line))
+                for line in _lines:
+                    if id(line) not in used:
+                        ordered_lines.append(line)
+                _order = [_lines.index(l) for l in ordered_lines]
+            else:
+                _order = None
         else:
             _order = None
 
@@ -377,8 +462,6 @@ def segment(im: PIL.Image.Image,
     else:
         script_detection = False
 
-    # create objects and assign IDs
-    blls = []
     _shp_regs = {}
     for reg_type, rgs in regions.items():
         for reg in rgs:
@@ -386,16 +469,19 @@ def segment(im: PIL.Image.Image,
 
     # reorder lines
     logger.debug(f'Reordering baselines with main RO function {reading_order_fn}.')
-    basic_lo = reading_order_fn(lines=lines, regions=_shp_regs.values(), text_direction=text_direction[-2:])
+    all_regions = [reg for rgs in regions.values() for reg in rgs]
+    basic_lo = reading_order_fn(lines=lines, regions=all_regions, text_direction=text_direction[-2:])
     lines = [lines[idx] for idx in basic_lo]
 
+    # assign region IDs
+    blls = []
     for line in lines:
         line_regs = []
         for reg_id, reg in _shp_regs.items():
-            line_ls = geom.LineString(line['baseline'])
+            line_ls = geom.LineString(line.baseline)
             if is_in_region(line_ls, reg):
                 line_regs.append(reg_id)
-        blls.append(BaselineLine(id=f'_{uuid.uuid4()}', baseline=line['baseline'], boundary=line['boundary'], tags=line['tags'], regions=line_regs))
+        blls.append(replace(line, regions=line_regs))
 
     return Segmentation(text_direction=text_direction,
                         imagename=getattr(im, 'filename', None),
