@@ -18,12 +18,14 @@ Utility functions for data loading and training of VGSL networks.
 import json
 import numbers
 from collections import Counter
+from collections.abc import Sequence
 from functools import partial
-from typing import Any, Dict, List, Sequence, Tuple, Union
+from typing import Any, Union
 
 import torch
 import torch.nn.functional as F
 from torchvision import transforms
+from torchvision.transforms import v2
 from importlib import resources
 
 from kraken.lib import functional_im_transforms as F_t
@@ -39,7 +41,7 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-def _get_type(tags: Dict, default='default') -> str:
+def _get_type(tags: dict, default='default') -> str:
     if tags is None:
         return default
     ot = tags.get('type', [{'type': default}])[0]
@@ -55,9 +57,10 @@ class ImageInputTransforms(transforms.Compose):
                  height: int,
                  width: int,
                  channels: int,
-                 pad: Union[int, Tuple[int, int], Tuple[int, int, int, int]],
+                 pad: Union[int, tuple[int, int], tuple[int, int, int, int]],
                  valid_norm: bool = True,
-                 force_binarization: bool = False) -> None:
+                 force_binarization: bool = False,
+                 dtype: torch.dtype = torch.float32) -> None:
         """
         Container for image input transforms for recognition and segmentation
         networks.
@@ -73,15 +76,17 @@ class ImageInputTransforms(transforms.Compose):
                         standard scaling.
             force_binarization: Forces binarization of input images using the
                                 nlbin algorithm.
+            dtype: Data type of the output tensors.
         """
         super().__init__(None)
 
-        self._scale: Tuple[int, int] = (height, width)
+        self._scale: tuple[int, int] = (height, width)
         self._valid_norm = valid_norm
         self._force_binarization = force_binarization
         self._batch = batch
         self._channels = channels
-        self.pad = pad
+        self._pad = pad
+        self._dtype = dtype
 
         self._create_transforms()
 
@@ -118,23 +123,42 @@ class ImageInputTransforms(transforms.Compose):
                                        'combination with forced binarization.')
 
         self.transforms = []
-        self.transforms.append(transforms.Lambda(partial(F_t.pil_to_mode, mode=self._mode)))
+
+        def mode_transform():
+            if self._mode == 'L':
+                return v2.Grayscale(num_output_channels=1)
+            return partial(F_t.pil_to_mode, mode=self._mode)
+
+        def resize_transform(im):
+            oh, ow = self._scale
+            h, w = im.height, im.width
+            if oh == 0:
+                oh = int(h * ow / w)
+            elif ow == 0:
+                ow = int(w * oh / h)
+            return v2.functional.resize(im, [oh, ow], interpolation=v2.InterpolationMode.LANCZOS, antialias=True)
+
+        self.transforms.append(mode_transform())
 
         if self._force_binarization:
-            self.transforms.append(transforms.Lambda(F_t.pil_to_bin))
+            self.transforms.append(F_t.pil_to_bin)
         if self._scale != (0, 0):
             if self._center_norm:
                 lnorm = CenterNormalizer(self._scale[0])
-                self.transforms.append(transforms.Lambda(partial(F_t.pil_dewarp, lnorm=lnorm)))
-                self.transforms.append(transforms.Lambda(partial(F_t.pil_to_mode, mode=self._mode)))
+                self.transforms.append(partial(F_t.pil_dewarp, lnorm=lnorm))
+                self.transforms.append(mode_transform())
             else:
-                self.transforms.append(transforms.Lambda(partial(F_t.pil_fixed_resize, scale=self._scale)))
+                if self._scale[0] > 0 and self._scale[1] > 0:
+                    self.transforms.append(v2.Resize(self._scale, interpolation=v2.InterpolationMode.LANCZOS, antialias=True))
+                else:
+                    self.transforms.append(resize_transform)
         if self._pad:
-            self.transforms.append(transforms.Pad(self._pad, fill=255))
-        self.transforms.append(transforms.ToTensor())
+            self.transforms.append(v2.Pad(self._pad, fill=255))
+        self.transforms.append(v2.PILToTensor())
+        self.transforms.append(v2.ToDtype(self._dtype, scale=True))
         # invert
-        self.transforms.append(transforms.Lambda(F_t.tensor_invert))
-        self.transforms.append(transforms.Lambda(partial(F_t.tensor_permute, perm=perm)))
+        self.transforms.append(v2.RandomInvert(p=1.0))
+        self.transforms.append(partial(F_t.tensor_permute, perm=perm))
 
     @property
     def batch(self) -> int:
@@ -201,7 +225,7 @@ class ImageInputTransforms(transforms.Compose):
         return self._mode if not self.force_binarization else '1'
 
     @property
-    def scale(self) -> Tuple[int, int]:
+    def scale(self) -> tuple[int, int]:
         """
         Desired output shape (height, width) of the image. If any value is set
         to 0, image will be rescaled proportionally with height, width, if 1
@@ -214,7 +238,7 @@ class ImageInputTransforms(transforms.Compose):
             return self._scale
 
     @scale.setter
-    def scale(self, scale: Tuple[int, int]) -> None:
+    def scale(self, scale: tuple[int, int]) -> None:
         self._scale = scale
         self._create_transforms()
 
@@ -226,7 +250,7 @@ class ImageInputTransforms(transforms.Compose):
         return self._pad
 
     @pad.setter
-    def pad(self, pad: Union[int, Tuple[int, int], Tuple[int, int, int, int]]) -> None:
+    def pad(self, pad: Union[int, tuple[int, int], tuple[int, int, int, int]]) -> None:
         if not isinstance(pad, (numbers.Number, tuple, list)):
             raise TypeError('Got inappropriate padding arg')
         self._pad = pad
@@ -266,7 +290,7 @@ class ImageInputTransforms(transforms.Compose):
         self._create_transforms()
 
 
-def global_align(seq1: Sequence[Any], seq2: Sequence[Any]) -> Tuple[int, List[str], List[str]]:
+def global_align(seq1: Sequence[Any], seq2: Sequence[Any]) -> tuple[int, list[str], list[str]]:
     """
     Computes a global alignment of two strings.
 
@@ -296,8 +320,8 @@ def global_align(seq1: Sequence[Any], seq2: Sequence[Any]) -> Tuple[int, List[st
             direction[i][j] = best[0]
     d = cost[-1][-1]
     # backtrace
-    algn1: List[Any] = []
-    algn2: List[Any] = []
+    algn1: list[Any] = []
+    algn2: list[Any] = []
     i = len(direction) - 1
     j = len(direction[0]) - 1
     while direction[i][j] != (-1, 0):
@@ -329,7 +353,7 @@ def compute_confusions(algn1: Sequence[str], algn2: Sequence[str]):
         insertions, `del` an integer of the number of deletions, `subs` per
         script substitutions.
     """
-    counts: Dict[Tuple[str, str], int] = Counter()
+    counts: dict[tuple[str, str], int] = Counter()
     ref = resources.files('kraken.lib.dataset').joinpath('scripts.json')
     with ref.open('rb') as fp:
         script_map = json.load(fp)
@@ -340,10 +364,10 @@ def compute_confusions(algn1: Sequence[str], algn2: Sequence[str]):
                 return n
         return 'Unknown'
 
-    scripts: Dict[Tuple[str, str], int] = Counter()
-    ins: Dict[Tuple[str, str], int] = Counter()
+    scripts: dict[tuple[str, str], int] = Counter()
+    ins: dict[tuple[str, str], int] = Counter()
     dels: int = 0
-    subs: Dict[Tuple[str, str], int] = Counter()
+    subs: dict[tuple[str, str], int] = Counter()
     for u, v in zip(algn1, algn2):
         counts[(u, v)] += 1
     for k, v in counts.items():
