@@ -19,8 +19,7 @@ Adapted from:
 """
 import logging
 from dataclasses import dataclass, field
-from typing import (TYPE_CHECKING, Any, Dict, List, Literal, Optional,
-                    Sequence, Union)
+from typing import TYPE_CHECKING, Literal, Optional, Union
 
 import numpy as np
 import lightning as L
@@ -31,31 +30,31 @@ from torch.optim import lr_scheduler
 from torch.utils.data import DataLoader, Subset
 from torchmetrics.aggregation import MeanMetric
 
-from kraken.lib import default_specs
+from kraken.configs import ROTrainingConfig, ROTrainingDataConfig
 from kraken.lib.dataset import PageWiseROSet, PairWiseROSet
-from kraken.lib.ro.layers import MLP
+from kraken.lib.ro.layers import ROMLP
 from kraken.lib.segmentation import _greedy_order_decoder
-from kraken.lib.train import _configure_optimizer_and_lr_scheduler
+from kraken.train.utils import configure_optimizer_and_lr_scheduler
 
 if TYPE_CHECKING:
     from os import PathLike
-
     from torch.nn import Module
+    from kraken.models import BaseModel
 
 logger = logging.getLogger(__name__)
 
 
 @dataclass
 class DummyVGSLModel:
-    hyper_params: Dict[str, int] = field(default_factory=dict)
-    user_metadata: Dict[str, List] = field(default_factory=dict)
+    hyper_params: dict[str, int] = field(default_factory=dict)
+    user_metadata: dict[str, list] = field(default_factory=dict)
     one_channel_mode: Literal['1', 'L'] = '1'
     ptl_module: 'Module' = None
-    model_type: str = 'unknown'
+    model_type: list[str] = field(default_factory=list)
 
     def __post_init__(self):
-        self.hyper_params: Dict[str, int] = {'completed_epochs': 0}
-        self.user_metadata: Dict[str, List] = {'accuracy': [], 'metrics': []}
+        self.hyper_params: dict[str, int] = {'completed_epochs': 0}
+        self.user_metadata: dict[str, list] = {'accuracy': [], 'metrics': []}
 
     def save_model(self, filename):
         self.ptl_module.save_checkpoint(filename)
@@ -67,18 +66,7 @@ def spearman_footrule_distance(s, t):
 
 class RODataModule(L.LightningDataModule):
     def __init__(self,
-                 batch_size: int = 16000,
-                 training_data: Union[Sequence[Union['PathLike', str]], Sequence[Dict[str, Any]]] = None,
-                 evaluation_data: Optional[Union[Sequence[Union['PathLike', str]], Sequence[Dict[str, Any]]]] = None,
-                 partition: Optional[float] = 0.9,
-                 num_workers: int = 1,
-                 format_type: Literal['alto', 'page', 'xml'] = 'xml',
-                 level: Literal['baselines', 'regions'] = 'baselines',
-                 class_mapping: Optional[Dict[str, int]] = None,
-                 valid_entities: Optional[Sequence[str]] = None,
-                 merge_entities: Optional[Dict[str, str]] = None,
-                 merge_all_entities: bool = False,
-                 reading_order: Optional[str] = None):
+                 data_config: ROTrainingDataConfig):
         super().__init__()
 
         self.save_hyperparameters()
@@ -87,39 +75,27 @@ class RODataModule(L.LightningDataModule):
         """
         Builds the datasets.
         """
-        training_data = self.hparams.training_data
-        evaluation_data = self.hparams.evaluation_data
-        partition = self.hparams.partition
-
-        valid_entities = self.hparams.valid_entities
-        merge_entities = self.hparams.merge_entities
-        merge_all_entities = self.hparams.merge_all_entities
-
-        if not valid_entities:
-            valid_entities = None
+        training_data = self.hparams.data_config.training_data
+        evaluation_data = self.hparams.data_config.evaluation_data
+        partition = self.hparams.data_config.partition
 
         if not evaluation_data:
             np.random.shuffle(training_data)
             training_data = training_data[:int(partition*len(training_data))]
             evaluation_data = training_data[int(partition*len(training_data)):]
         train_set = PairWiseROSet(training_data,
-                                  mode=self.hparams.format_type,
-                                  level=self.hparams.level,
-                                  ro_id=self.hparams.reading_order,
-                                  class_mapping=self.hparams.class_mapping,
-                                  valid_entities=valid_entities,
-                                  merge_entities=merge_entities,
-                                  merge_all_entities=merge_all_entities)
+                                  mode=self.hparams.data_config.format_type,
+                                  level=self.hparams.data_config.level,
+                                  ro_id=self.hparams.data_config.reading_order,
+                                  class_mapping=self.hparams.data_config.class_mapping)
         self.train_set = Subset(train_set, range(len(train_set)))
         self.class_mapping = train_set.class_mapping
+        self.hparams.data_config.class_mapping = dict(self.class_mapping)
         val_set = PageWiseROSet(evaluation_data,
-                                mode=self.hparams.format_type,
+                                mode=self.hparams.data_config.format_type,
                                 class_mapping=self.class_mapping,
-                                level=self.hparams.level,
-                                ro_id=self.hparams.reading_order,
-                                valid_entities=valid_entities,
-                                merge_entities=merge_entities,
-                                merge_all_entities=merge_all_entities)
+                                level=self.hparams.data_config.level,
+                                ro_id=self.hparams.data_config.reading_order)
 
         self.val_set = Subset(val_set, range(len(val_set)))
 
@@ -132,8 +108,8 @@ class RODataModule(L.LightningDataModule):
 
     def train_dataloader(self):
         return DataLoader(self.train_set,
-                          batch_size=self.hparams.batch_size,
-                          num_workers=self.hparams.num_workers,
+                          batch_size=self.hparams.data_config.batch_size,
+                          num_workers=self.hparams.data_config.num_workers,
                           pin_memory=True,
                           shuffle=False)
 
@@ -141,7 +117,7 @@ class RODataModule(L.LightningDataModule):
         return DataLoader(self.val_set,
                           shuffle=False,
                           batch_size=1,
-                          num_workers=self.hparams.num_workers,
+                          num_workers=self.hparams.data_config.num_workers,
                           pin_memory=True)
 
     def get_feature_dim(self):
@@ -153,10 +129,8 @@ class RODataModule(L.LightningDataModule):
 
 class ROModel(L.LightningModule):
     def __init__(self,
-                 feature_dim: int,
-                 class_mapping: Dict[str, int],
-                 hyper_params: Dict[str, Any] = None,
-                 output: 'PathLike' = 'model') -> None:
+                 config: ROTrainingConfig,
+                 model: Optional['BaseModel'] = None) -> None:
         """
         A LightningModule encapsulating the unsupervised pretraining setup for
         a text recognition model.
@@ -173,36 +147,40 @@ class ROModel(L.LightningModule):
             **kwargs: Setup parameters, i.e. CLI parameters of the train() command.
         """
         super().__init__()
-        self.hyper_params = default_specs.READING_ORDER_HYPER_PARAMS
-        if hyper_params:
-            self.hyper_params.update(hyper_params)
+        self.save_hyperparameters(ignore=['model'])
 
-        self.output = output
+        if not isinstance(config, ROTrainingConfig):
+            raise ValueError(f'config attribute is {type(config)} not ROTrainingConfig.')
+
+        if model:
+            self.net = model
+
         self.criterion = torch.nn.BCEWithLogitsLoss()
-
-        self.best_epoch = -1
-        self.best_metric = torch.inf
-
-        logger.info('Creating new RO model')
-        self.ro_net = MLP(feature_dim, feature_dim * 2)
-
-        if 'file_system' in torch.multiprocessing.get_all_sharing_strategies():
-            logger.debug('Setting multiprocessing tensor sharing strategy to file_system')
-            torch.multiprocessing.set_sharing_strategy('file_system')
 
         self.nn = DummyVGSLModel(ptl_module=self)
 
         self.val_losses = MeanMetric()
         self.val_spearman = MeanMetric()
 
-        self.save_hyperparameters()
+    @classmethod
+    def load_from_weights(cls,
+                          path: Union[str, 'PathLike'],
+                          config: ROTrainingConfig) -> 'ROModel':
+        """
+        Initializes the module from a model weights file.
+        """
+        from kraken.models import load_models
+        models = load_models(path, tasks=['reading_order'])
+        if len(models) != 1:
+            raise ValueError(f'Found {len(models)} reading order models in model file.')
+        return cls(config=config, model=models[0])
 
     def forward(self, x):
-        return F.sigmoid(self.ro_net(x))
+        return F.sigmoid(self.net(x))
 
     def validation_step(self, batch, batch_idx):
         xs, ys, num_lines = batch['sample'], batch['target'], batch['num_lines']
-        logits = self.ro_net(xs).squeeze()
+        logits = self.net(xs).squeeze()
         yhat = F.sigmoid(logits)
         order = torch.zeros((num_lines, num_lines))
         idx = 0
@@ -223,10 +201,6 @@ class ROModel(L.LightningModule):
             val_metric = self.val_spearman.compute()
             val_loss = self.val_losses.compute()
 
-            if val_metric < self.best_metric:
-                logger.debug(f'Updating best metric from {self.best_metric} ({self.best_epoch}) to {val_metric} ({self.current_epoch})')
-                self.best_epoch = self.current_epoch
-                self.best_metric = val_metric
             logger.info(f'validation run: val_spearman {val_metric} val_loss {val_loss}')
             self.log('val_spearman', val_metric, on_step=False, on_epoch=True, prog_bar=True, logger=True)
             self.log('val_metric', val_metric, on_step=False, on_epoch=True, prog_bar=False, logger=True)
@@ -237,7 +211,7 @@ class ROModel(L.LightningModule):
 
     def training_step(self, batch, batch_idx):
         x, y = batch['sample'], batch['target']
-        logits = self.ro_net(x)
+        logits = self.net(x)
         loss = self.criterion(logits.squeeze(), y)
         self.log('loss',
                  loss,
@@ -247,18 +221,60 @@ class ROModel(L.LightningModule):
                  logger=True)
         return loss
 
-    def save_checkpoint(self, filename):
-        self.trainer.save_checkpoint(filename)
+    def on_load_checkpoint(self, checkpoint):
+        """
+        Reconstruct the model here and not in setup() as otherwise the weight
+        loading will fail.
+        """
+        from kraken.models import create_model
+        if not isinstance(checkpoint['hyper_parameters']['config'], ROTrainingConfig):
+            raise ValueError('Checkpoint is not a reading order model.')
+        data_config = checkpoint['datamodule_hyper_parameters']['data_config']
+        self.net = create_model('ROMLP',
+                                class_mapping=data_config.class_mapping,
+                                level=data_config.level)
+
+    def on_save_checkpoint(self, checkpoint):
+        """
+        Save hyperparameters a second time so we can set parameters that
+        shouldn't be overwritten in on_load_checkpoint.
+        """
+        checkpoint['_module_config'] = self.hparams.config
+
+    def setup(self, stage: Optional[str] = None):
+        if stage in [None, 'fit']:
+            set_class_mapping = dict(self.trainer.datamodule.get_class_mapping())
+            if not getattr(self, 'net', None):
+                logger.info('Creating new RO model')
+                self.net = ROMLP(class_mapping=set_class_mapping,
+                                 level=self.trainer.datamodule.hparams.data_config.level)
+            else:
+                net_class_mapping = self.net.user_metadata['class_mapping']
+                if set_class_mapping.keys() != net_class_mapping.keys():
+                    diff = set(set_class_mapping.keys()).symmetric_difference(set(net_class_mapping.keys()))
+                    raise ValueError(f'Training data and model class mapping differ: {diff}')
+                # backfill train_set/val_set mapping if key-equal as the actual
+                # numbering in the train_set might be different
+                num_classes = max(0, *net_class_mapping.values()) + 1
+                self.trainer.datamodule.train_set.dataset.class_mapping = net_class_mapping
+                self.trainer.datamodule.train_set.dataset.num_classes = num_classes
+                self.trainer.datamodule.val_set.dataset.class_mapping = net_class_mapping
+                self.trainer.datamodule.val_set.dataset.num_classes = num_classes
+
+            # store canonical (one-to-one) class mapping in model metadata for inference
+            self.net.user_metadata['class_mapping'] = self.trainer.datamodule.train_set.dataset.canonical_class_mapping
+
+            logger.info('Training types:')
+            for k, v in set_class_mapping.items():
+                logger.info(f'  {k}\t{v}\t{self.trainer.datamodule.train_set.dataset.class_stats[k]}')
 
     def configure_callbacks(self):
         callbacks = []
-        if self.hparams.hyper_params['quit'] == 'early':
+        if self.hparams.config.quit == 'early':
             callbacks.append(EarlyStopping(monitor='val_metric',
                                            mode='min',
-                                           patience=self.hparams.hyper_params['lag'],
+                                           patience=self.hparams.config.lag,
                                            stopping_threshold=0.0))
-        if self.hparams.hyper_params['pl_logger']:
-            callbacks.append(LearningRateMonitor(logging_interval='step'))
         return callbacks
 
     # configuration of optimizers and learning rate schedulers
@@ -268,9 +284,9 @@ class ROModel(L.LightningModule):
     # batch-wise learning rate warmup. In lr_scheduler_step() calls to the
     # scheduler are then only performed at the end of the epoch.
     def configure_optimizers(self):
-        return _configure_optimizer_and_lr_scheduler(self.hparams.hyper_params,
-                                                     self.ro_net.parameters(),
-                                                     loss_tracking_mode='min')
+        return configure_optimizer_and_lr_scheduler(self.hparams.config,
+                                                    self.net.parameters(),
+                                                    loss_tracking_mode='min')
 
     def optimizer_step(self, epoch, batch_idx, optimizer, optimizer_closure):
         # update params
@@ -278,13 +294,13 @@ class ROModel(L.LightningModule):
 
         # linear warmup between 0 and the initial learning rate `lrate` in `warmup`
         # steps.
-        if self.hparams.hyper_params['warmup'] and self.trainer.global_step < self.hparams.hyper_params['warmup']:
-            lr_scale = min(1.0, float(self.trainer.global_step + 1) / self.hparams.hyper_params['warmup'])
+        if self.hparams.config.warmup and self.trainer.global_step < self.hparams.config.warmup:
+            lr_scale = min(1.0, float(self.trainer.global_step + 1) / self.hparams.config.warmup)
             for pg in optimizer.param_groups:
-                pg["lr"] = lr_scale * self.hparams.hyper_params['lrate']
+                pg["lr"] = lr_scale * self.hparams.config.lrate
 
     def lr_scheduler_step(self, scheduler, metric):
-        if not self.hparams.hyper_params['warmup'] or self.trainer.global_step >= self.hparams.hyper_params['warmup']:
+        if not self.hparams.config.warmup or self.trainer.global_step >= self.hparams.config.warmup:
             # step OneCycleLR each batch if not in warmup phase
             if isinstance(scheduler, lr_scheduler.OneCycleLR):
                 scheduler.step()
