@@ -109,9 +109,9 @@ def build_binary_dataset(files: Optional[list[Union[str, 'PathLike', 'Segmentati
                          output_file: Union[str, 'PathLike'] = None,
                          format_type: Literal['xml', 'alto', 'page', None] = 'xml',
                          num_workers: int = 0,
-                         ignore_splits: bool = False,
+                         ignore_splits: bool = True,
                          random_split: Optional[tuple[float, float, float]] = None,
-                         force_type: Optional[str] = None,
+                         force_type: Optional[Literal['kraken_recognition_bbox', 'kraken_recognition_baseline']] = None,
                          recordbatch_size: int = 100,
                          skip_empty_lines: bool = True,
                          callback: Callable[[int, int], None] = lambda chunk, lines: None,
