@@ -15,5 +15,5 @@
 """
 Tools for trainable reading order.
 """
-
 from .model import ROModel, RODataModule  # NOQA
+from .layers import ROMLP  # NOQA
