@@ -15,7 +15,13 @@ resources = thisfile / 'resources'
 
 class TestModels(unittest.TestCase):
     """
-    Testing model loading routines
+    Testing legacy model loading routines (`kraken.lib.models.load_any`).
+
+    .. deprecated::
+        These tests exercise the deprecated `kraken.lib.models.load_any` API
+        which will be removed with kraken 8. New code should use
+        `kraken.models.load_models` instead. See `test_plugins.py` for tests
+        of the replacement loading API.
     """
 
     def setUp(self):
