@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-import json
+import copy
+import pickle
 import tempfile
 import unittest
 import uuid
@@ -73,13 +74,14 @@ class TestSerializations(unittest.TestCase):
     Tests for output serialization
     """
     def setUp(self):
-        with open(resources /'records.json', 'r') as fp:
-            self.box_records = [containers.BBoxOCRRecord(**x) for x in json.load(fp)]
+        with open(resources / 'box_rec.pkl', 'rb') as fp:
+            box_seg = pickle.load(fp)
+        self.box_records = box_seg.lines
 
-        with open(resources / 'bl_records.json', 'r') as fp:
-            recs = json.load(fp)
-            self.bl_records = [containers.BaselineOCRRecord(**bl) for bl in recs['lines']]
-            self.bl_regions = recs['regions']
+        with open(resources / 'bl_rec.pkl', 'rb') as fp:
+            bl_seg = pickle.load(fp)
+        self.bl_records = bl_seg.lines
+        self.bl_regions = bl_seg.regions
 
         self.box_segmentation = containers.Segmentation(type='bbox',
                                                         imagename='foo.png',
@@ -88,10 +90,14 @@ class TestSerializations(unittest.TestCase):
                                                         script_detection=True,
                                                         regions={})
 
+        bl_records_no_regions = copy.deepcopy(self.bl_records)
+        for line in bl_records_no_regions:
+            line.regions = []
+
         self.bl_segmentation = containers.Segmentation(type='baselines',
                                                        imagename='foo.png',
                                                        text_direction='horizontal-lr',
-                                                       lines=self.bl_records,
+                                                       lines=bl_records_no_regions,
                                                        script_detection=True,
                                                        regions={})
 
@@ -124,15 +130,13 @@ class TestSerializations(unittest.TestCase):
                                                                    'models': 'bar.mlmodel',
                                                                    'pad': 16,
                                                                    'bidi_reordering': True})]
-
-
     def test_box_vertical_hocr_serialization(self):
         """
         Test vertical line hOCR serialization
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.box_segmentation, writing_mode='vertical-lr', template='hocr'))
+        fp.write(serialization.serialize(self.box_segmentation, image_size=(2000, 2000), writing_mode='vertical-lr', template='hocr'))
         validate_hocr(self, fp)
 
     def test_box_hocr_serialization(self):
@@ -141,7 +145,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.box_segmentation, template='hocr'))
+        fp.write(serialization.serialize(self.box_segmentation, image_size=(2000, 2000), template='hocr'))
         validate_hocr(self, fp)
 
     def test_box_alto_serialization_validation(self):
@@ -150,7 +154,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.box_segmentation, template='alto'))
+        fp.write(serialization.serialize(self.box_segmentation, image_size=(2000, 2000), template='alto'))
         validate_alto(self, fp)
 
     def test_box_abbyyxml_serialization_validation(self):
@@ -159,7 +163,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.box_segmentation, template='abbyyxml'))
+        fp.write(serialization.serialize(self.box_segmentation, image_size=(2000, 2000), template='abbyyxml'))
         doc = etree.fromstring(fp.getvalue().encode('utf-8'))
         with open(resources / 'FineReader10-schema-v1.xml') as schema_fp:
             abbyy_schema = etree.XMLSchema(etree.parse(schema_fp))
@@ -171,7 +175,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.box_segmentation, template='pagexml'))
+        fp.write(serialization.serialize(self.box_segmentation, image_size=(2000, 2000), template='pagexml'))
         validate_page(self, fp)
 
     def test_bl_alto_serialization_validation(self):
@@ -180,7 +184,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation, template='alto'))
+        fp.write(serialization.serialize(self.bl_segmentation, image_size=(2000, 2000), template='alto'))
         validate_alto(self, fp)
         roundtrip(self, self.bl_records, fp)
 
@@ -190,7 +194,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation, template='abbyyxml'))
+        fp.write(serialization.serialize(self.bl_segmentation, image_size=(2000, 2000), template='abbyyxml'))
         doc = etree.fromstring(fp.getvalue().encode('utf-8'))
         with open(resources / 'FineReader10-schema-v1.xml') as schema_fp:
             abbyy_schema = etree.XMLSchema(etree.parse(schema_fp))
@@ -202,7 +206,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation, template='pagexml'))
+        fp.write(serialization.serialize(self.bl_segmentation, image_size=(2000, 2000), template='pagexml'))
         validate_page(self, fp)
         roundtrip(self, self.bl_records, fp)
 
@@ -212,7 +216,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation_regs, template='alto'))
+        fp.write(serialization.serialize(self.bl_segmentation_regs, image_size=(2000, 2000), template='alto'))
         validate_alto(self, fp)
         roundtrip(self, self.bl_records, fp)
 
@@ -222,7 +226,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation_regs, template='abbyyxml'))
+        fp.write(serialization.serialize(self.bl_segmentation_regs, image_size=(2000, 2000), template='abbyyxml'))
         doc = etree.fromstring(fp.getvalue().encode('utf-8'))
         with open(resources / 'FineReader10-schema-v1.xml') as schema_fp:
             abbyy_schema = etree.XMLSchema(etree.parse(schema_fp))
@@ -234,7 +238,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation_regs, template='pagexml'))
+        fp.write(serialization.serialize(self.bl_segmentation_regs, image_size=(2000, 2000), template='pagexml'))
         validate_page(self, fp)
         roundtrip(self, self.bl_records, fp)
 
@@ -244,7 +248,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_seg_nolines_regs, template='alto'))
+        fp.write(serialization.serialize(self.bl_seg_nolines_regs, image_size=(2000, 2000), template='alto'))
         validate_alto(self, fp)
 
     def test_region_only_abbyyxml_serialization_validation(self):
@@ -253,7 +257,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_seg_nolines_regs, template='abbyyxml'))
+        fp.write(serialization.serialize(self.bl_seg_nolines_regs, image_size=(2000, 2000), template='abbyyxml'))
         doc = etree.fromstring(fp.getvalue().encode('utf-8'))
         with open(resources / 'FineReader10-schema-v1.xml') as schema_fp:
             abbyy_schema = etree.XMLSchema(etree.parse(schema_fp))
@@ -265,7 +269,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_seg_nolines_regs, template='pagexml'))
+        fp.write(serialization.serialize(self.bl_seg_nolines_regs, image_size=(2000, 2000), template='pagexml'))
         validate_page(self, fp)
 
     def test_bl_region_alto_serialization_validation_steps(self):
@@ -274,7 +278,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation, template='alto', processing_steps=self.metadata_steps))
+        fp.write(serialization.serialize(self.bl_segmentation, image_size=(2000, 2000), template='alto', processing_steps=self.metadata_steps))
 
         validate_alto(self, fp)
         roundtrip(self, self.bl_records, fp)
@@ -285,7 +289,7 @@ class TestSerializations(unittest.TestCase):
         """
         fp = StringIO()
 
-        fp.write(serialization.serialize(self.bl_segmentation, template='abbyyxml', processing_steps=self.metadata_steps))
+        fp.write(serialization.serialize(self.bl_segmentation, image_size=(2000, 2000), template='abbyyxml', processing_steps=self.metadata_steps))
 
         doc = etree.fromstring(fp.getvalue().encode('utf-8'))
         with open(resources / 'FineReader10-schema-v1.xml') as schema_fp:
@@ -297,7 +301,6 @@ class TestSerializations(unittest.TestCase):
         Validates output with processing steps against PageXML schema
         """
         fp = StringIO()
-        fp.write(serialization.serialize(self.bl_segmentation, template='pagexml', processing_steps=self.metadata_steps))
+        fp.write(serialization.serialize(self.bl_segmentation, image_size=(2000, 2000), template='pagexml', processing_steps=self.metadata_steps))
         validate_page(self, fp)
         roundtrip(self, self.bl_records, fp)
-
